Refactor addedTransaction middleware to transactions middleware
