+++ src/store/middlewares/index.js
@@ -2,7 +2,7 @@ import thunk from 'redux-thunk';
 import metronomeMiddleware from './metronome';
 import accountMiddleware from './account';
 import loginMiddleware from './login';
-import addedTransactionMiddleware from './addedTransaction';
+import transactionsMiddleware from './transactions';
 import loadingBarMiddleware from './loadingBar';
 import offlineMiddleware from './offline';
 import notificationMiddleware from './notification';
@@ -11,7 +11,7 @@ import savedAccountsMiddleware from './savedAccounts';
 
 export default [
   thunk,
-  addedTransactionMiddleware,
+  transactionsMiddleware,
   loginMiddleware,
   metronomeMiddleware,
   accountMiddleware,
